Moving database settings to separate file
